+++ integration/src/test/java/org/mule/test/streaming/PayloadStatisticsTestCase.java
@@ -20,8 +20,6 @@ import static org.mule.test.allure.AllureConstants.StreamingFeature.STREAMING;
 import static org.mule.test.allure.AllureConstants.StreamingFeature.StreamingStory.STATISTICS;
 
 import org.mule.runtime.api.exception.MuleException;
-import org.mule.runtime.api.exception.MuleRuntimeException;
-import org.mule.runtime.api.lifecycle.Disposable;
 import org.mule.runtime.core.api.construct.Flow;
 import org.mule.runtime.core.api.management.stats.PayloadStatistics;
 import org.mule.runtime.http.api.HttpService;
@@ -32,9 +30,6 @@ import org.mule.runtime.http.api.domain.message.response.HttpResponse;
 import org.mule.service.http.TestHttpClient;
 import org.mule.tck.junit4.rule.DynamicPort;
 import org.mule.tck.junit4.rule.SystemProperty;
-import org.mule.tck.probe.JUnitLambdaProbe;
-import org.mule.tck.probe.PollingProber;
-import org.mule.tck.probe.Probe;
 import org.mule.test.AbstractIntegrationTestCase;
 import org.mule.test.runner.RunnerDelegateTo;
 import org.mule.tests.api.TestQueueManager;
@@ -66,8 +61,6 @@ public class PayloadStatisticsTestCase extends AbstractIntegrationTestCase {
 
   public static final int BYTES_SIZE = 1343;
 
-  public static final int MUTANT_SUMMON_BYTE_SIZE = 84;
-
   @ClassRule
   public static TemporaryFolder temporaryFolder = new TemporaryFolder();
 
@@ -92,18 +85,6 @@ public class PayloadStatisticsTestCase extends AbstractIntegrationTestCase {
   @Inject
   @Named("listOfMessagesSource")
   public Flow listOfMessagesSource;
-
-  @Inject
-  @Named("streamSource")
-  public Flow streamSource;
-
-  @Inject
-  @Named("iteratorSource")
-  public Flow iteratorSource;
-
-  @Inject
-  @Named("iteratorSourceConsumeOnResponse")
-  public Flow iteratorSourceConsumeOnResponse;
 
   private final String configFile;
 
@@ -269,74 +250,6 @@ public class PayloadStatisticsTestCase extends AbstractIntegrationTestCase {
     assertThat(fileListStatistics.getInputByteCount(), is(0L));
     assertThat(fileListStatistics.getOutputObjectCount(), is(6L));
     assertThat(fileListStatistics.getOutputByteCount(), is(0L));
-  }
-
-  @Test
-  @Description("Assert statistics for an source that generates an InputStream")
-  public void streamSource() throws Exception {
-    streamSource.start();
-
-    queueManager.read("streamSourceComplete", RECEIVE_TIMEOUT, MILLISECONDS).getMessage();
-
-    final PayloadStatistics fileListStatistics =
-        muleContext.getStatistics().getPayloadStatistics("streamSource/source");
-
-    assertThat(fileListStatistics.getComponentIdentifier(), is("marvel:magneto-mutant-summon"));
-
-    assertThat(fileListStatistics.getInvocationCount(), is(1L));
-
-    assertThat(fileListStatistics.getInputObjectCount(), is(0L));
-    // do not count the container message
-    assertThat(fileListStatistics.getOutputObjectCount(), is(0L));
-    assertThat(fileListStatistics.getOutputByteCount(), is(MUTANT_SUMMON_BYTE_SIZE * 1L));
-
-    new PollingProber().check(new JUnitLambdaProbe(() -> {
-      assertThat(fileListStatistics.getInputByteCount(), is(MUTANT_SUMMON_BYTE_SIZE * 1L));
-      return true;
-    }));
-  }
-
-  @Test
-  @Description("Assert statistics for a source that returns an Iterator")
-  public void iteratorSource() throws Exception {
-    iteratorSource.start();
-
-    queueManager.read("iteratorSourceComplete", RECEIVE_TIMEOUT, MILLISECONDS).getMessage();
-
-    final PayloadStatistics fileListStatistics =
-        muleContext.getStatistics().getPayloadStatistics("iteratorSource/source");
-
-    assertThat(fileListStatistics.getComponentIdentifier(), is("marvel:magneto-brotherhood"));
-
-    assertThat(fileListStatistics.getInvocationCount(), is(1L));
-
-    assertThat(fileListStatistics.getInputObjectCount(), is(0L));
-    assertThat(fileListStatistics.getInputByteCount(), is(0L));
-    assertThat(fileListStatistics.getOutputObjectCount(), is(6L));
-    assertThat(fileListStatistics.getOutputByteCount(), is(0L));
-  }
-
-  @Test
-  @Description("Assert statistics for a source that consumes an iterator on response")
-  public void iteratorSourceConsumeElementsOnResponse() throws Exception {
-    iteratorSourceConsumeOnResponse.start();
-
-    queueManager.read("iteratorSourceConsumeOnResponseComplete", RECEIVE_TIMEOUT, MILLISECONDS).getMessage();
-
-    final PayloadStatistics fileListStatistics =
-        muleContext.getStatistics().getPayloadStatistics("iteratorSourceConsumeOnResponse/source");
-
-    assertThat(fileListStatistics.getComponentIdentifier(), is("marvel:magneto-brotherhood"));
-
-    assertThat(fileListStatistics.getInvocationCount(), is(1L));
-    assertThat(fileListStatistics.getInputByteCount(), is(0L));
-    assertThat(fileListStatistics.getOutputObjectCount(), is(0L));
-    assertThat(fileListStatistics.getOutputByteCount(), is(0L));
-
-    new PollingProber().check(new JUnitLambdaProbe(() -> {
-      assertThat(fileListStatistics.getInputObjectCount(), is(6L));
-      return true;
-    }));
   }
 
 }
