+++ integration/src/test/resources/org/mule/streaming/payload-statistics-non-repeatable-config.xml
@@ -18,15 +18,6 @@
     <test-components:queue-config name="listOfMessagesSourceComplete">
         <test-components:connection />
     </test-components:queue-config>
-    <test-components:queue-config name="streamSourceComplete">
-        <test-components:connection />
-    </test-components:queue-config>
-    <test-components:queue-config name="iteratorSourceComplete">
-        <test-components:connection />
-    </test-components:queue-config>
-    <test-components:queue-config name="iteratorSourceConsumeOnResponseComplete">
-        <test-components:connection />
-    </test-components:queue-config>
 
     <http:listener-config name="simpleConf">
         <http:listener-connection host="localhost" port="${port}"/>
@@ -54,18 +45,6 @@
         <flow-ref name="consumeStreamsOnce"/>
 
         <test-components:queue-push config-ref="listOfMessagesSourceComplete"/>
-    </flow>
-    
-    <flow name="streamSource" initialState="stopped">
-        <marvel:magneto-mutant-summon>
-            <marvel:response>
-                <marvel:body>#[payload]</marvel:body>
-            </marvel:response>
-        </marvel:magneto-mutant-summon>
-    
-        <flow-ref name="consumeStreamOnce"/>
-        
-        <test-components:queue-push config-ref="streamSourceComplete"/>
     </flow>
 
     <flow name="listOfMessagesOperation">
@@ -106,24 +85,6 @@
         </marvel:wolverine-blacklist>
     
         <flow-ref name="consumeStreamsOnce"/>
-    </flow>
-    
-   <flow name="iteratorSource" initialState="stopped">
-        <marvel:magneto-brotherhood />
-    
-        <flow-ref name="consumeStreamsOnce"/>
-        
-        <test-components:queue-push config-ref="iteratorSourceComplete"/>
-    </flow>
-    
-      <flow name="iteratorSourceConsumeOnResponse" initialState="stopped">
-        <marvel:magneto-brotherhood>
-            <marvel:response>
-                <marvel:body>#[payload]</marvel:body>
-            </marvel:response>
-        </marvel:magneto-brotherhood>
-    
-        <test-components:queue-push config-ref="iteratorSourceConsumeOnResponseComplete"/>
     </flow>
 
     <!-- Utility sub-flow's -->
