+++ integration/src/test/resources/org/mule/streaming/payload-statistics-source-config.xml
@@ -1,74 +1,58 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <mule xmlns="http://www.mulesoft.org/schema/mule/core"
       xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-      xmlns:http="http://www.mulesoft.org/schema/mule/http"
-      xmlns:file="http://www.mulesoft.org/schema/mule/file"
       xmlns:marvel="http://www.mulesoft.org/schema/mule/marvel"
       xmlns:test-components="http://www.mulesoft.org/schema/mule/test-components"
       xsi:schemaLocation="
        http://www.mulesoft.org/schema/mule/core http://www.mulesoft.org/schema/mule/core/current/mule.xsd
-       http://www.mulesoft.org/schema/mule/file http://www.mulesoft.org/schema/mule/file/current/mule-file.xsd
-       http://www.mulesoft.org/schema/mule/http http://www.mulesoft.org/schema/mule/http/current/mule-http.xsd
        http://www.mulesoft.org/schema/mule/marvel http://www.mulesoft.org/schema/mule/marvel/current/mule-marvel.xsd
        http://www.mulesoft.org/schema/mule/test-components http://www.mulesoft.org/schema/mule/test-components/current/mule-test-components.xsd">
 
     <test-components:queue-config name="queue">
         <test-components:connection />
     </test-components:queue-config>
-    <test-components:queue-config name="listOfMessagesSourceComplete">
+    <test-components:queue-config name="streamSourceComplete">
         <test-components:connection />
     </test-components:queue-config>
-
-    <http:listener-config name="simpleConf">
-        <http:listener-connection host="localhost" port="${port}"/>
-    </http:listener-config>
-
-    <file:config name="file">
-        <file:connection workingDir="${workingDir}" />
-    </file:config>
-
-    <flow name="bytesSource">
-        <http:listener path="/" config-ref="simpleConf"/>
+    <test-components:queue-config name="iteratorSourceComplete">
+        <test-components:connection />
+    </test-components:queue-config>
+    <test-components:queue-config name="iteratorSourceConsumeOnResponseComplete">
+        <test-components:connection />
+    </test-components:queue-config>
+    
+    <flow name="streamSource" initialState="stopped">
+        <marvel:magneto-mutant-summon>
+            <marvel:response>
+                <marvel:body>#[payload]</marvel:body>
+            </marvel:response>
+        </marvel:magneto-mutant-summon>
     
         <flow-ref name="consumeStream"/>
+        
+        <test-components:queue-push config-ref="streamSourceComplete"/>
     </flow>
 
-    <flow name="listOfMessagesSource" initialState="stopped">
-        <marvel:cerebro-detect-new-mutants/>
+    <flow name="iteratorSource" initialState="stopped">
+        <marvel:magneto-brotherhood>
+            <marvel:response>
+                <marvel:body>#[payload]</marvel:body>
+            </marvel:response>
+        </marvel:magneto-brotherhood>
     
         <flow-ref name="consumeStreams"/>
-
-        <test-components:queue-push config-ref="listOfMessagesSourceComplete"/>
+        
+        <test-components:queue-push config-ref="iteratorSourceComplete"/>
     </flow>
-
-    <flow name="listOfMessagesOperation">
-        <marvel:adamantium-injectors injectorsToCreate="4" injectorLoad="${bytesSize}"/>
     
-        <flow-ref name="consumeStreams"/>
-    </flow>
-
-    <flow name="pagesOfMessagesOperation">
-        <file:list config-ref="file" directoryPath="#[vars.path]"/>
+    <flow name="iteratorSourceConsumeOnResponse" initialState="stopped">
+        <marvel:magneto-brotherhood>
+            <marvel:response>
+                <marvel:body>#[payload]</marvel:body>
+            </marvel:response>
+        </marvel:magneto-brotherhood>
     
-        <flow-ref name="consumeStreams"/>
-    </flow>
-
-    <flow name="pagedOperation">
-        <marvel:get-relics/>
-    
-        <flow-ref name="consumeStreams"/>
-    </flow>
-
-    <flow name="streamOperation">
-        <marvel:to-stream/>
-    
-        <flow-ref name="consumeStream"/>
-    </flow>
-
-    <flow name="iteratorOperation">
-        <marvel:wolverine-blacklist/>
-    
-        <flow-ref name="consumeStreams"/>
+        <test-components:queue-push config-ref="iteratorSourceConsumeOnResponseComplete"/>
     </flow>
 
 <!--     Utility sub-flow's -->
